+++ app/src/main/java/com/example/buggyweather/home/presenter/HomeViewModel.kt
@@ -7,9 +7,9 @@ import com.example.buggyweather.core.base.BaseViewModel
 import com.example.buggyweather.core.base.UseCase
 import com.example.buggyweather.domain.CurrentWeather
 import com.example.buggyweather.domain.MeasuringUnits
-import com.example.buggyweather.network.exception.BadRequestException
-import com.example.buggyweather.network.exception.NotFoundException
-import com.example.buggyweather.network.exception.RemoteException
+import com.example.buggyweather.core.network.exception.BadRequestException
+import com.example.buggyweather.core.network.exception.NotFoundException
+import com.example.buggyweather.core.network.exception.RemoteException
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
