+++ app/src/main/java/com/example/buggyweather/home/repository/CurrentWeatherService.kt
@@ -2,7 +2,7 @@ package com.example.buggyweather.home.repository
 
 import com.example.buggyweather.domain.CurrentWeather
 import com.example.buggyweather.domain.MeasuringUnits
-import com.example.buggyweather.utils.Constants
+import com.example.buggyweather.core.utils.Constants
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
