--- app/src/main/java/com/example/buggyweather/network/exception/BadRequestException.kt
@@ -1,3 +1,0 @@
-package com.example.buggyweather.network.exception
-
-class BadRequestException(msg: String): Throwable(msg)
