--- app/src/main/java/com/example/buggyweather/network/exception/NotFoundException.kt
@@ -1,3 +1,0 @@
-package com.example.buggyweather.network.exception
-
-class NotFoundException(msg: String): Throwable(msg)
