+++ app/src/test/java/com/example/buggyweather/home/usecase/GetCurrentWeatherUseCaseTest.kt
@@ -5,8 +5,8 @@ import com.example.buggyweather.domain.CurrentWeather
 import com.example.buggyweather.domain.MeasuringUnits
 import com.example.buggyweather.helper.BaseTest
 import com.example.buggyweather.home.repository.CurrentWeatherDataSource
-import com.example.buggyweather.network.KnownExceptionMessage
-import com.example.buggyweather.network.exception.NotFoundException
+import com.example.buggyweather.core.network.KnownExceptionMessage
+import com.example.buggyweather.core.network.exception.NotFoundException
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.doThrow
 import com.nhaarman.mockitokotlin2.verify
