+++ app/src/test/java/com/example/buggyweather/whole/repository/WholeDayWeatherMapperTest.kt
@@ -3,7 +3,7 @@ package com.example.buggyweather.whole.repository
 import androidx.arch.core.util.Function
 import com.example.buggyweather.domain.WholeDayWeather
 import com.example.buggyweather.helper.BaseTest
-import com.example.buggyweather.network.exception.RemoteException
+import com.example.buggyweather.core.network.exception.RemoteException
 import okhttp3.ResponseBody.Companion.toResponseBody
 import org.junit.Assert
 import org.junit.Before
