+++ app/src/test/java/com/example/buggyweather/whole/usecase/GetWholeDayWeatherUseCaseTest.kt
@@ -6,8 +6,8 @@ import com.example.buggyweather.domain.ForecastRequest
 import com.example.buggyweather.domain.MeasuringUnits
 import com.example.buggyweather.domain.WholeDayWeather
 import com.example.buggyweather.helper.BaseTest
-import com.example.buggyweather.network.KnownExceptionMessage
-import com.example.buggyweather.network.exception.NotFoundException
+import com.example.buggyweather.core.network.KnownExceptionMessage
+import com.example.buggyweather.core.network.exception.NotFoundException
 import com.example.buggyweather.whole.repository.WholeDayWeatherDataSource
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.doThrow
