+++ core/src/main/java/com/example/buggyweather/core/network/exception/BadRequestException.kt
@@ -1,0 +1,3 @@
+package com.example.buggyweather.core.network.exception
+
+class BadRequestException(msg: String): Throwable(msg)
