+++ core/src/main/java/com/example/buggyweather/core/network/exception/NotFoundException.kt
@@ -1,0 +1,3 @@
+package com.example.buggyweather.core.network.exception
+
+class NotFoundException(msg: String): Throwable(msg)
