+++ core/src/main/java/com/example/buggyweather/core/network/exception/RemoteException.kt
@@ -1,3 +1,3 @@
-package com.example.buggyweather.network.exception
+package com.example.buggyweather.core.network.exception
 
 class RemoteException(msg: String): Throwable("Server error: $msg")
