+++ core/src/main/java/com/example/buggyweather/core/utils/Constants.kt
@@ -1,4 +1,4 @@
-package com.example.buggyweather.utils
+package com.example.buggyweather.core.utils
 
 object Constants {
 	const val BASE_URL = "https://api.openweathermap.org/data/2.5/"
