+++ core/src/main/java/com/example/buggyweather/core/utils/IntExtension.kt
@@ -1,4 +1,4 @@
-package com.example.buggyweather.utils
+package com.example.buggyweather.core.utils
 
 import android.icu.text.SimpleDateFormat
